+++ test/utils.ts
@@ -1,0 +1,1 @@
+export const req = (path: string) => new Request(`http://localhost${path}`)